BUG/MINOR: mux_quic: convert QCC rx.rlen to 64bits
This patch is related to the issue reported on the previous issue
related to QMux record length parsing.

QCC rx.rlen is used to store the decoded record length. Convert it into
a plain 64bits integer instead of a size_t. This ensures it is
sufficient to decode record length, even with an increase of
max_record_length value (not currently implemented).

This should fix github build issue #3334 for 32bits architecture.

No need to backport.

+++ include/haproxy/mux_quic-t.h
@@ -93,7 +93,7 @@ struct qcc {
 	} tx;
 	struct {
 		struct buffer qstrm_buf;
-		size_t rlen; /* last record length read */
+		uint64_t rlen; /* last record length read */
 	} rx;
 
 	uint64_t largest_bidi_r; /* largest remote bidi stream ID opened. */
